api/requirements.txt
fastapi
uvicorn[standard]
kubernetes
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
cryptography
python-multipart
openpyxl
